requirements.txt
PyQt5
pynput
keyboard
Pillow
python-telegram-bot
pywin32
packaging
requests
numpy